Add localStorage load to functions.js module
